Ticket [https://trac.osgeo.org/fdo/ticket/475]: Build: Fix Warnings and Compile Errors
git-svn-id: http://svn.osgeo.org/fdo/trunk@4426 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Utilities/ExpressionEngine/Src/Functions/Date/FdoFunctionCurrentDate.cpp
@@ -143,7 +143,7 @@ FdoLiteralValue *FdoFunctionCurrentDate::Evaluate (
     dt.day     = local_time.tm_mday;
     dt.hour    = local_time.tm_hour;
     dt.minute  = local_time.tm_min;
-    dt.seconds = local_time.tm_sec;
+    dt.seconds = (float)local_time.tm_sec;
 
     // Return the requested information.
